PerformanceResource Program restricted

+++ backend/stout/src/main/java/edu/mtech/stout/resources/PerformanceResource.java
@@ -1,8 +1,13 @@
 package edu.mtech.stout.resources;
 
+import edu.mtech.stout.api.QueryBySelector;
 import edu.mtech.stout.api.Status;
 import edu.mtech.stout.core.Performance;
+import edu.mtech.stout.core.Program;
+import edu.mtech.stout.core.User;
 import edu.mtech.stout.db.PerformanceDAO;
+import edu.mtech.stout.db.ProgramDAO;
+import io.dropwizard.auth.Auth;
 import io.dropwizard.hibernate.UnitOfWork;
 import io.dropwizard.jersey.PATCH;
 import io.dropwizard.jersey.params.LongParam;
@@ -10,6 +15,7 @@ import io.dropwizard.jersey.params.LongParam;
 import javax.annotation.security.PermitAll;
 import javax.annotation.security.RolesAllowed;
 import javax.ws.rs.*;
+import java.util.List;
 
 
 @Path("/performance-indicators/{performanceId}")
@@ -17,17 +23,37 @@ import javax.ws.rs.*;
 public class PerformanceResource {
 
   PerformanceDAO dao;
+  ProgramDAO programDao;
+  QueryBySelector queryBySelector;
 
-  public PerformanceResource(PerformanceDAO dao){
+  public PerformanceResource(PerformanceDAO dao, ProgramDAO programDao){
+
     this.dao = dao;
+    this.programDao = programDao;
+    queryBySelector = new QueryBySelector(programDao);
   }
 
   @GET
   @PermitAll
   @UnitOfWork
-  public Performance getPerformance(@PathParam("performanceId")LongParam performanceId){
-    
-    return findSafely(performanceId.get());
+  public Performance getPerformance(@Auth User user, @PathParam("performanceId")LongParam performanceId){
+    List<Program> p = programDao.findByPerformance(performanceId.get());
+    if(p.size() > 0){
+      Boolean hasAccess = false;
+      for(int i = 0; i < p.size(); i++){
+        if(queryBySelector.queryByProgramId(user, p.get(i).getId() )){
+          hasAccess = true;
+        }
+      }
+      if(hasAccess){
+        return findSafely(performanceId.get());
+      }else{
+        throw new NotAuthorizedException("Cannot get a performance indicator not in your program");
+      }
+    }else{
+      throw new NotFoundException("The requested performance indicator does not exist");
+    }
+
   }
 
   private Performance findSafely(long performanceId){
@@ -36,31 +62,59 @@ public class PerformanceResource {
 
   @PATCH
   @UnitOfWork
-  public Performance updatePeformance(@PathParam("performanceId") LongParam performanceId, Performance performance){
-    return dao.update(performance);
+  public Performance updatePeformance(@Auth User user, @PathParam("performanceId") LongParam performanceId, Performance performance){
+    List<Program> p = programDao.findByPerformance(performanceId.get());
+    if(p.size() > 0){
+      Boolean hasAccess = false;
+      for(int i = 0; i < p.size(); i++){
+        if(queryBySelector.queryByProgramId(user, p.get(i).getId() )){
+          hasAccess = true;
+        }
+      }
+      if(hasAccess){
+        return dao.update(performance);
+      }else{
+        throw new NotAuthorizedException("Cannot get a performance indicator not in your program");
+      }
+    }else{
+      throw new NotFoundException("The requested performance indicator does not exist");
+    }
   }
 
   @DELETE
   @RolesAllowed({"Program Coordinator"})
   @UnitOfWork
-  public Status deletePerformance(@PathParam("performanceId") LongParam performanceId) {
-    Status status = new Status();
-    status.setId(performanceId.get().intValue());
-    status.setAction("DELETE");
-    status.setResource("Performance");
+  public Status deletePerformance(@Auth User user, @PathParam("performanceId") LongParam performanceId) {
+    List<Program> p = programDao.findByPerformance(performanceId.get());
+    if(p.size() > 0){
+      Boolean hasAccess = false;
+      for(int i = 0; i < p.size(); i++){
+        if(queryBySelector.queryByProgramId(user, p.get(i).getId() )){
+          hasAccess = true;
+        }
+      }
+      if(hasAccess){
+        Status status = new Status();
+        status.setId(performanceId.get().intValue());
+        status.setAction("DELETE");
+        status.setResource("Performance");
 
-    boolean success = dao.delete(findSafely(performanceId.get().intValue()));
+        boolean success = dao.delete(findSafely(performanceId.get().intValue()));
 
-    if (success) {
-      status.setMessage("Successfully deleted Performance Indicator");
-      status.setStatus(200);
-    } else {
-      status.setMessage("Error deleting Performance Indicator");
-      status.setStatus(500);
+        if (success) {
+          status.setMessage("Successfully deleted Performance Indicator");
+          status.setStatus(200);
+        } else {
+          status.setMessage("Error deleting Performance Indicator");
+          status.setStatus(500);
+        }
+
+        return status;
+      }else{
+        throw new NotAuthorizedException("Cannot get a performance indicator not in your program");
+      }
+    }else{
+      throw new NotFoundException("The requested performance indicator does not exist");
     }
-
-    return status;
-
   }
-
 }
